src/participant/requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
package-name==0.1
Pillow==10.1.0
pycryptodome==3.19.0
pypng==0.20220715.0
qrcode==7.4.2
requests==2.31.0
typing_extensions==4.8.0
urllib3==2.1.0
Werkzeug==3.0.1
